+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+  - 🌟 [@mixte/components] 新增 `Scrollbar` 组件 ( 开发中 )
   - 💄 [@mixte/components] 优化 `GridTable` 加载中蒙版层级实现逻辑 *
 
 ## [v3.5.0-beta.8]